requirements.txt
python==3.9.0

# 核心数据处理
pandas==2.0.3
numpy==1.26.0

# Word文档处理
python-docx==0.8.11

# 正则表达式
regex==2023.10.3

# Windows COM接口
pywin32==306

# 并行计算
joblib==1.3.2

# 机器学习/深度学习 (CUDA 12.1)
torch==2.1.0+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
scikit-learn==1.3.2

# 中文NLP（LTP）
ltp==4.2.14

# 隐私脱敏
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# 类型提示
typing_extensions==4.8.0